refactor: move mock_supervisor to conftest at tests
