+++ agents/tests/conftest.py
@@ -2,6 +2,7 @@ import logging
 import os
 import tempfile
 
+import mock
 import pytest
 
 
@@ -63,3 +64,15 @@ class=logging.Formatter
 
     os.remove(config_file.name)
     os.remove(log_file.name)
+
+
+@pytest.fixture
+def mock_supervisor(setup_temp_file_logging):
+    from agents.supervisor import AgentSupervisor
+
+    agent = mock.MagicMock()
+    agent.server_name = 'mock-server'
+
+    supervisor = AgentSupervisor(agent)
+
+    return supervisor
